Removed obsolete components, save issue persists
